Re-wrote Joe keymap of top of interact
The Joe mode is quite broken (but so it was before the change).
Regression: Search/Replace in not implemented (but is was broken in core anyway)
I belive this is quite harmless, Joe is not a primary keymap anyway.
In any case, that change makes Yi.MakeKeymap obsolete; allowing for further
improvement of the Yi.Interact module, which is to become the common base
for all keymaps; easing the development of all keymaps.
(eg. interactive search/replace will benefit from support in Interact)