Disentangle UI/Core/Editor dependencies
This is a broad patch, that cleans up the general architecture of Yi.
Now the architecture can be faithfully described as a series of layers,
each cleanly living on top of each other:

Keymap Code       (Emacs.hs, Vim.hs, etc.)
Keymap primitives (Core.hs, Keymap.hs)
UI                (gtk, vty)
Editor            (Editor.hs = stack of buffers + other editor state)
Buffer            (Buffer.hs)
Buffer primitives (FastBuffer.hs)

There are more local simplifications to do, but this is the bulk of the work,
so I record as it is.