Adjusted Yi.Keymap.Emacs.Utils.gotoLineE to accept an optional column number as well
It is also more forgiving, before anything other than digits would cause yi to error out.
Now we just print an error message to the status line. The line and column number is
pretty forgiving essentially you just need to separate the two numbers by any string
not containing digits. So "3 this is the separator 4" will work just as well as "3.4
or "3:4" etc.
Finally this does mean that the goto line feature of the emacs key bindings differs
from the emacs one, but since it still does the same thing on normal input
(ie just a line number) I think it is not a bad thing.