Fixed Yi.Buffer.Normal.execB for line up and down movements.
Basically now if you are on the first line instead of doing nothing you move
to the start of the line (ie the start of the buffer). Similarly if you
are on the last line of the file you move to the end of line
(ie the end of the buffer). This is much nicer to work with.
While I was at it I've added cleaned up the emacs keybinding definitions
for keys and used Yi.Buffer.Normal.execB in a couple of places.