Fixed the indentation to the last opening bracket.
Before it did not take into the closing brackets and now it does,
such that if we open a bracket but close it on the same line then
the opening bracket is not returned as an indentation. We still
may return a bracket further back if that one is unmatched.