Support for increasing and decreasing indentation styles
I have added support to generalise the Yi.Indent.autoIndentHelperB
with a style argument. The style argument can (currently) be one of the
following:
 IncreaseCycle
 IncreaseOnly
 DecreaseCycle
 DecreaseOnly
with the intuitive semantics for each one based on the current indentation
and a list of indentation hints.
This is in preparation to purge issue number 110.
Things remaining: I don't know how to specify Shift-Tab as a key.
Also I have defined 'adjIndent' which is essentially a wrapper for
creating a YiM action using the given indentation behaviour argument (one of the above)
together with the current mode's indentation function. Currently this function
resides in Yi/Keymap/Emacs/Utils.hs but I think it should go elsewhere (in particular
it is not limited to use with Emacs keybindings) but I couldn't decide on the most
suitable place for it.