UNDO: Fix for issue number 117 (switch using substring of buffer name)
I've generalised "Yi.Editor.switchToBufferWithNameE" to accept a predicate
over buffer names and will switch to the first buffer which satisfies this
predicate. To retain proper behaviour for the empty string, if the empty string
satisfies the predicate then we switch to the next buffer.
Arguably we could split this functionality and have a
"Yi.Editor.switchToBufferSatisfiesNameE" and then the above function is just
a specialisation.