Vim mode moves to matching char with '%'
Unfortunately, the current '%' leaves this always in an ambiguous state. The percentage-movement '%' should only be active once a number has been inputted; otherwise, this '%' function should always run.

I'm not sure how to modify the movement '%' currently, because any function that takes an integer argument only takes it /optionally/, other than this one (it seems). Vim only moves if an integer is provided, and otherwise treats it like the matching movement action.