Fix scrolling in the Pango UI
Currently the amount moved is hard-coded to 10 lines. Before, it tried to use the winh value (divided by two), but this never functioned correctly due to winh's odd behaviour. It seems to me that each keymap might want to handle scrolling past the edge of the window on its own.