Extract darwin-specific executeFile behavior
executeFile cannot be used under Darwin directly as it raises
ENOTSUP when called from a multithreaded process. The workaround
is to call forkProcess first. This patch extracts the workaround
from buildLaunch and makes it available to restart as well.
In addition, it adds code to stop the current process when
the child process is stopped. This makes UI.suspend work properly
in the Vty UI.

Note that while this does re-enable the reloadEditor command under
Darwin, that command is still not usable as this executeFile
does not stop any background threads, including Vty's input-consuming
thread. This means that after reloading, both the old and the new
processes are still consuming input, and the result is quite unusable.