JS: A bunch of changes (read long desc)
This patch changes the stroking functionality to no longer use Foldable and
instead uses "manual" stroking.  This could probably be done with generics of
some kind in the future.

I'm also introducing the Failable class which is used to represent "error
constructors" for datatypes.  This allows us to use a generic "please" function
instead of specific ones for different datatypes.

It also contains a new function called "hate" which is basically a weighted
recoverWith function which allows more fine-grained control over precedence.

Furthermore, I changed the relationship between statements and expressions. Not
all expressions are valid statements anymore, even though "real" JavaScript has
this.  This is a compromise to help performance, which was pretty bad due to
ambiguity between blocks and JavaScript objects.