Let the window region be a mere value
This simplifies things for what's coming.
Also, this probably speeds up things, since the region is calculated 
only once per layout. When it's a function from FBuffer, it must be
recomputed every time the region is fetched.
(And I'm just about to add a systematic usage of the region field right
after the layout)